update: 結合具合が疎なコマンドを個別のファイルにした
