+++ src-tauri/Cargo.toml
@@ -13,16 +13,7 @@ edition = "2021"
 tauri-build = { version = "1.2", features = [] }
 
 [dependencies]
-tauri = { version = "1.2", features = [
-    "app-all",
-    "dialog-all",
-    "os-all",
-    "path-all",
-    "shell-open",
-    "window-close",
-    "window-set-title",
-    "window-start-dragging",
-] }
+tauri = { version = "1.2", features = ["app-all", "dialog-all", "os-all", "path-all", "shell-open", "window-close", "window-set-always-on-top", "window-set-title", "window-start-dragging"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 [target.'cfg(target_os = "linux")'.dependencies]
